my-service/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
dateparser==1.1.8
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
pycryptodome==3.19.0
python-binance==1.0.19
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
six==1.16.0
tzlocal==5.0.1
ujson==5.8.0
urllib3<2.0
websockets==11.0.3
yarl==1.9.2
